% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalwebsrv-package.R
\name{available_sources}
\alias{available_sources}
\title{GDAL web source names}
\usage{
available_sources()
}
\value{
character vector
}
\description{
Names of available image servers
}
\examples{
available_sources()
}
