% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalwebsrv.R
\docType{data}
\name{gdal_web_sources}
\alias{gdal_web_sources}
\title{GDAL web servers}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 2 columns.
}
\usage{
gdal_web_sources
}
\description{
Source files for various image servers available via GDAL.
}
\details{
\code{gdal_web_sources} is a data frame of \code{name}, and \code{file} which is relative to the installed
package (see Examples).
}
\examples{
system.file(gdal_web_sources$file[1L], package = "gdalwebsrv", mustWork = TRUE)
}
\keyword{dataset}
