% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalwebsrv-package.R, R/gdalwebsrv.R
\docType{package}
\name{gdalwebsrv-package}
\alias{gdalwebsrv}
\alias{gdalwebsrv-package}
\title{gdalwebsrv: Web Services for Imagery}
\description{
Provides some sources for imagery from web servers.

Provides files from the GDAL project to access online imagery servers.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
