% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalwebsrv-package.R
\name{lerc_file}
\alias{lerc_file}
\title{GDAL terrain source (LERC)}
\usage{
lerc_file()
}
\value{
character string file path
}
\description{
Find file providing ESRI server for terrain, LERC.
}
\examples{
lerc_file()
}
