% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalwebsrv-package.R
\name{server_file}
\alias{server_file}
\title{GDAL web sources files}
\usage{
server_file(name)
}
\arguments{
\item{name}{name of source}
}
\value{
character vector
}
\description{
Find files providinging image server by name. These files may be used to access
raster data using functions from packages raster, stars, vapour, terra, sf (and others) or
from other tools that are GDAL enabled.
}
\details{
See \code{\link[=available_sources]{available_sources()}} for available server names.
}
\examples{
server_file(gdal_web_sources$name[1L])
}
